+++ output.h
@@ -1,5 +1,5 @@
 void output_init();
-long long get_time();
+uint64_t get_time();
 void output_write_command(char* cmd, char* name, int salary);
 void output_lock_status(char* status);
 void output_condition_variables(char* signal);
